[sg2525] fix: fixed emulator makefile, and added a 1MB memory check for literal instructions

+++ src/assembler/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 #include <stdint.h>
-#define MB (1<<20)
+#define MB (int32_t)(1<<20)
 #define XZR 31
 
 typedef enum {EQ=0, NE = 1, GE = 10, LT = 11, GT = 12, LE = 13, AL = 14} conds;
